Test: Update RecaptchaEnterpriseCallbackTest to use defaultJourney and remove unused code
